SPORTS-53 Move redirecting logic to files in hocs folder
