+++ frontend/src/components/hocs/redirectIfAuthenticated.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Redirect } from 'react-router'
+import { branch, renderComponent } from 'recompose';
+
+const checkAuthentication = (({ isAuthenticated }) => isAuthenticated);
+
+const RedirectHere = ({ location, isAuthenticated }) => {
+  console.log('from another', location);
+  const newRoute = location.state && location.state.from || { pathname: '/' };
+
+  return <Redirect to={newRoute} />;
+
+}
+const redirectIfAuthenticated = branch(
+  checkAuthentication,
+  renderComponent(RedirectHere)
+);
+
+export default redirectIfAuthenticated;
