+++ frontend/src/constants/routes.js
@@ -1,3 +1,4 @@
+export const ROOT = '/';
 export const ADMIN = '/admin';
 export const LOGIN = '/login';
 export const TOURNAMENT_TREE = '/tournament/tree';
