+++ frontend/src/utils/getAuthDetails.js
@@ -3,6 +3,7 @@ import { LOCAL_AUTH_VARIABLE } from '../constants/constants';
 export default function getAuthDetails() {
   try {
     let authDetails = JSON.parse(localStorage.getItem(LOCAL_AUTH_VARIABLE));
+    
     if (authDetails instanceof Object && 'isAuthenticated' in authDetails && authDetails.refreshToken) {
       return authDetails;
     }
